+++ CHANGES.md
@@ -34,6 +34,13 @@
 
 ### New Modules
 
+  * `XMonad.Util.MutexScratchpads`
+
+    Named scratchpads that can be mutually exclusive: This new module extends the
+    idea of named scratchpads such that you can define "families of scratchpads"
+    that are mutually exclusive on the same screen. It also allows to remove this
+    constraint of being mutually exclusive with another scratchpad when.
+
   * `XMonad.Hooks.Focus`
 
     A new module extending ManageHook EDSL to work on focused windows and